information shown will change depending on type
depending on the type of information the status bar is suppose to display it will change based on its appearance. Along with that the stats it will show will either be a percentage or number of cases. This depends on whether it is the number of active cases: cases, or recovered/mortality rate: %.