testing if api call returns an empty json
added methods to process the data received from the api call: (processData method)
note: when using class methods in react/javascript must use this keyword before method name to use it within the own class.
added an if else statement in the event the api call returns an empty json object, because of this you can call another fetch request within it and process the data further.