made it a grid to allow for different layouts depending on screens size
changed the display to a grid and centered the items according to the grid in order for it to be responsive on different screen sizes.
For ipad/tablet screen size need to give it a width so it can be centered with even spacing between the 2 columns and edges of the screen.